Improved PS Solving
Added gradient and hessian for solving price schedule problem. Much more stable. Works with b <= 40. Beyond that it becomes numerically unstable due to small values of the distribution. Perhaps solvable by forcing everything into big float for that part of the problem.

Constrained moment inequality estimator to focus on b<=40 with a = 1. It wants to put everything at the bound, but I think that's defensible.