Changes to optimal_price_schedule
optimal_price_schedule now returns a dictionary. Keys are integers representing N (as in N part tariff) and values are the corresponding optimal price schedule objects.

This was done to minimize computation. For example, with an observed value of N=5, a call to optimal_price_schedule with N=6 will generate all necessary price schedules to compare forgone profits and calculate bounds on complexity cost. This is an improvement as previously 2 calls would be needed (N=5, N=6) which calculates duplicate price schedules (N=2,3,4,5).

Changes are reflected in LIquorQD_tests but not other files yet.